NoSQL Scattered Database solution for the .Net framework. A highly scaleable combination of a Key/Value and a Document database.
Store nearly anything instantly and easily, without the need of a
database server. All you need to access to the folder where the data is
stored. This is great for tiny applications, or giant enterprise
applications because it scales from a single machine almost to infinity,
with almost no degradation in performance from the first record to the
trillionth. You can have as many nodes as can share a file (almost
limitless), as many records as disk can hold (this grows every year), as
much Sharding of data as folder structures allow (again, almost
limitless). From a scaling perspective, it is very robust. The very
granular, performant approach used removes almost all limits to size,
accessibility and availability. The limits are controlled by the
infrastructure and not by ZarahDB. If the OS can reach the file, ZarahDB
can work with it.